// src/app/shared/servicios/productos/model/producto-model.ts
export class Product{
 id_producto: string;
 codigo: string;
 nombre: string;
 descripcion: string;
 precio_compra: number;
 precio_venta: number;
 firma_farmaceutica: string;
 concentracion: string;
 registro_sanitario: string;
 estado: boolean;
 transaccion: string;
 fecha_transaccion: string;
 usuario: string;
}